Do not log the popen output, propagate change for logger.
The output from popen when we call the graph extractor can be huge. We
do not want to store that in memory (through the communicate method)
and we do not want to store it in the log specific for a repository.

So, now we do not log the popen output anymore but we redirect it to
stdout (we may have a better solution to redirect that output to a
file).